requirements.txt
fastapi>=0.135.1
langchain-core>=0.2.0
langchain-postgres>=0.0.6
mcp>=1.26.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
requests>=2.32.5
uvicorn>=0.27.0

# W3C traceparent propagation (только API, без SDK и span-экспорта)
opentelemetry-api>=1.20.0
